Add prettier as a commit hook, switch to eslint flat config, add some tests surrounding deleting relationships (#139)
I believe most/all of the non prettier changes are package.json,
pnpm-lock, eslint configs,
and some tests in imodeltransformerhub.test.ts:
1) it("should be able to handle relationship delete using new
relationship provenance method with no fedguids"
1) it("should be able to handle relationship delete using fedguids"
1) it("should be able to handle relationship delete using old
relationship provenance method with no fedguids"

---------

+++ packages/performance-tests/test/rawInserts.ts
@@ -1,5 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import { ChangesetFileProps } from "@itwin/core-common";
-import { Element, ElementGroupsMembers, IModelDb, SnapshotDb, StandaloneDb } from "@itwin/core-backend";
+import {
+  Element,
+  ElementGroupsMembers,
+  IModelDb,
+  SnapshotDb,
+  StandaloneDb,
+} from "@itwin/core-backend";
 import { IModelTransformer } from "@itwin/imodel-transformer";
 import { IModelTransformerTestUtils } from "@itwin/imodel-transformer/lib/cjs/test/IModelTransformerUtils";
 import { Logger, OpenMode } from "@itwin/core-bentley";
@@ -19,13 +29,19 @@ const iModelName = "Many PhysicalObjects and Relationships";
 const ELEM_COUNT = 100_000;
 assert(ELEM_COUNT % 2 === 0, "elem count must be divisible by 2");
 
-export default async function rawInserts(reporter: Reporter, branchName: string) {
-
+export default async function rawInserts(
+  reporter: Reporter,
+  branchName: string
+) {
   Logger.logInfo(loggerCategory, "starting 150k entity inserts");
 
   let testIModel: TestIModel | undefined;
   const [insertsTimer] = timed(() => {
-    testIModel = generateTestIModel({ numElements: 100_000, fedGuids: true, fileName: `RawInserts-source.bim` });
+    testIModel = generateTestIModel({
+      numElements: 100_000,
+      fedGuids: true,
+      fileName: `RawInserts-source.bim`,
+    });
   });
 
   if (testIModel === undefined)
@@ -39,21 +55,31 @@ export default async function rawInserts(reporter: Reporter, branchName: string)
     "time elapsed (seconds)",
     insertsTimer?.elapsedSeconds ?? -1,
     {
-      "Element Count": IModelTransformerTestUtils.count(sourceDb, Element.classFullName),
-      "Relationship Count": IModelTransformerTestUtils.count(sourceDb, ElementGroupsMembers.classFullName),
+      "Element Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        Element.classFullName
+      ),
+      "Relationship Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        ElementGroupsMembers.classFullName
+      ),
       "Branch Name": branchName,
     }
   );
 
   sourceDb.saveChanges();
 
-  Logger.logInfo(loggerCategory, "Done. Starting changeset application of same content");
+  Logger.logInfo(
+    loggerCategory,
+    "Done. Starting changeset application of same content"
+  );
 
   const changeset1 = createChangeset(sourceDb);
   const changesetDbPath = initOutputFile(`RawInsertsApply.bim`, outputDir);
-  if (fs.existsSync(changesetDbPath))
-    fs.unlinkSync(changesetDbPath);
-  const changesetDb = StandaloneDb.createEmpty(changesetDbPath, { rootSubject: { name: "RawInsertsApply" } });
+  if (fs.existsSync(changesetDbPath)) fs.unlinkSync(changesetDbPath);
+  const changesetDb = StandaloneDb.createEmpty(changesetDbPath, {
+    rootSubject: { name: "RawInsertsApply" },
+  });
 
   const [applyChangeSetTimer] = timed(() => {
     changesetDb.nativeDb.applyChangeset(changeset1);
@@ -65,17 +91,30 @@ export default async function rawInserts(reporter: Reporter, branchName: string)
     "time elapsed (seconds)",
     applyChangeSetTimer?.elapsedSeconds ?? -1,
     {
-      "Element Count": IModelTransformerTestUtils.count(sourceDb, Element.classFullName),
-      "Relationship Count": IModelTransformerTestUtils.count(sourceDb, ElementGroupsMembers.classFullName),
+      "Element Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        Element.classFullName
+      ),
+      "Relationship Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        ElementGroupsMembers.classFullName
+      ),
       "Branch Name": branchName,
     }
   );
 
-  Logger.logInfo(loggerCategory, "Done. Starting with-provenance transformation of same content");
+  Logger.logInfo(
+    loggerCategory,
+    "Done. Starting with-provenance transformation of same content"
+  );
 
   const targetPath = initOutputFile(`RawInserts-Target.bim`, outputDir);
-  const targetDb = SnapshotDb.createEmpty(targetPath, { rootSubject: { name: "RawInsertsTarget" } });
-  const transformerWithProv = new IModelTransformer(sourceDb, targetDb, { noProvenance: false });
+  const targetDb = SnapshotDb.createEmpty(targetPath, {
+    rootSubject: { name: "RawInsertsTarget" },
+  });
+  const transformerWithProv = new IModelTransformer(sourceDb, targetDb, {
+    noProvenance: false,
+  });
 
   const [transformWithProvTimer] = await timed(async () => {
     await transformerWithProv.processAll();
@@ -87,17 +126,33 @@ export default async function rawInserts(reporter: Reporter, branchName: string)
     "time elapsed (seconds)",
     transformWithProvTimer?.elapsedSeconds ?? -1,
     {
-      "Element Count": IModelTransformerTestUtils.count(sourceDb, Element.classFullName),
-      "Relationship Count": IModelTransformerTestUtils.count(sourceDb, ElementGroupsMembers.classFullName),
+      "Element Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        Element.classFullName
+      ),
+      "Relationship Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        ElementGroupsMembers.classFullName
+      ),
       "Branch Name": branchName,
     }
   );
 
-  Logger.logInfo(loggerCategory, "Done. Starting without-provenance transformation of same content");
+  Logger.logInfo(
+    loggerCategory,
+    "Done. Starting without-provenance transformation of same content"
+  );
 
-  const targetNoProvPath = initOutputFile(`RawInserts-TargetNoProv.bim`, outputDir);
-  const targetNoProvDb = SnapshotDb.createEmpty(targetNoProvPath, { rootSubject: { name: "RawInsertsTarget" } });
-  const transformerNoProv = new IModelTransformer(sourceDb, targetNoProvDb, { noProvenance: true });
+  const targetNoProvPath = initOutputFile(
+    `RawInserts-TargetNoProv.bim`,
+    outputDir
+  );
+  const targetNoProvDb = SnapshotDb.createEmpty(targetNoProvPath, {
+    rootSubject: { name: "RawInsertsTarget" },
+  });
+  const transformerNoProv = new IModelTransformer(sourceDb, targetNoProvDb, {
+    noProvenance: true,
+  });
 
   const [transformNoProvTimer] = await timed(async () => {
     await transformerNoProv.processAll();
@@ -109,8 +164,14 @@ export default async function rawInserts(reporter: Reporter, branchName: string)
     "time elapsed (seconds)",
     transformNoProvTimer?.elapsedSeconds ?? -1,
     {
-      "Element Count": IModelTransformerTestUtils.count(sourceDb, Element.classFullName),
-      "Relationship Count": IModelTransformerTestUtils.count(sourceDb, ElementGroupsMembers.classFullName),
+      "Element Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        Element.classFullName
+      ),
+      "Relationship Count": IModelTransformerTestUtils.count(
+        sourceDb,
+        ElementGroupsMembers.classFullName
+      ),
       "Branch Name": branchName,
     }
   );
